requirements.txt
customtkinter
ultralytics
pillow
numpy
opencv-python-headless
packaging
